Address review: harden DDP-ignore against private API drift, re-apply after PEFT wrap
- Wrap the private DistributedDataParallel._set_params_and_buffers_to_ignore_for_model
  call in try/except, falling back to setting _ddp_params_and_buffers_to_ignore
  directly so a future PyTorch API change can't block model loading.
- Move _exclude_rope_inv_freq_from_ddp to loader_utils.py (shared by loader.py,
  llama.py, vision.py without circular imports) and call it again after
  get_peft_model wraps the model in a PeftModel, since the rotary buffers'
  fully qualified names change once nested under "base_model.model...".
