+++ unsloth/models/loader.py
@@ -33,6 +33,7 @@ from transformers import AutoConfig
 from transformers import __version__ as transformers_version
 from peft import PeftConfig, PeftModel
 from .loader_utils import (
+    _exclude_rope_inv_freq_from_ddp,
     _get_fp8_mode_and_check_settings,
     _offline_quantize_to_fp8,
     _tag_model_with_fp8_torchao_config,
@@ -279,33 +280,6 @@ def _fix_rope_inv_freq(model):
                             module.multi_gpu_short_sin_cached[device_idx] = (
                                 emb.sin() * module.scaling_factor
                             ).to(dtype = dtype, device = device_obj, non_blocking = True)
-    return model
-
-
-# Rotary inv_freq buffers are deliberately kept on CPU - Unsloth pre-builds a
-# cos/sin cache per GPU instead (see LlamaRotaryEmbedding.multi_gpu_cos_cached)
-# so the GPU-resident lookup never needs to move the tiny inv_freq tensor itself.
-# torch.nn.parallel.DistributedDataParallel ignores device entirely when it
-# broadcasts buffers across ranks, so a CPU buffer crashes NCCL's
-# _broadcast_coalesced with "No backend type associated with device type cpu".
-# Telling DDP to skip these specific buffers avoids that crash without moving
-# inv_freq to GPU (which would break the per-GPU cache design) and without
-# disabling buffer broadcast for every other module (the user's workaround).
-# https://github.com/unslothai/unsloth/issues/6656
-_ROTARY_INV_FREQ_BUFFER_NAMES = ("inv_freq", "short_inv_freq", "long_inv_freq")
-
-
-def _exclude_rope_inv_freq_from_ddp(model):
-    ignored = list(getattr(model, "_ddp_params_and_buffers_to_ignore", None) or [])
-    for module_name, module in model.named_modules():
-        for buffer_name, _ in module.named_buffers(recurse = False):
-            if buffer_name in _ROTARY_INV_FREQ_BUFFER_NAMES:
-                fqn = f"{module_name}.{buffer_name}" if module_name else buffer_name
-                if fqn not in ignored:
-                    ignored.append(fqn)
-    if ignored:
-        from torch.nn.parallel import DistributedDataParallel
-        DistributedDataParallel._set_params_and_buffers_to_ignore_for_model(model, ignored)
     return model
